# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My DOC API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/profiles": {
            "get": {
                "tags": [
                    "Профиль"
                ],
                "summary": "Профиль текущего User",
                "operationId": "71f675beb4d242986837e8162cb38b31",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "Timmy"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Black"
                                        },
                                        "date_of_birth": {
                                            "type": "string",
                                            "format": "date",
                                            "example": null
                                        },
                                        "login": {
                                            "type": "string",
                                            "example": "Lok1"
                                        },
                                        "balance": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 167.2
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Профиль"
                ],
                "summary": "Обновление профиля у User",
                "operationId": "084f1de8bc082c118b5c206c6bbc7443",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "first_name": {
                                                "type": "string",
                                                "example": "Timmy"
                                            },
                                            "last_name": {
                                                "type": "string",
                                                "example": "Black"
                                            },
                                            "date_of_birth": {
                                                "type": "string",
                                                "format": "date",
                                                "example": null
                                            },
                                            "login": {
                                                "type": "string",
                                                "example": "Lok1"
                                            },
                                            "balance": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 167.2
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "first_name": {
                                            "type": "string",
                                            "example": "Timmy"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Tamminsky"
                                        },
                                        "date_of_birth": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "12.12.2002"
                                        },
                                        "login": {
                                            "type": "string",
                                            "example": "Lok1"
                                        },
                                        "balance": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 12.1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unautoruzed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/transactions": {
            "post": {
                "tags": [
                    "Транзакции"
                ],
                "summary": "Создание транзакции",
                "operationId": "a864ad0735fd71aa4e588151a00ced80",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "amount": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 10.2
                                            },
                                            "currency": {
                                                "type": "string",
                                                "example": "Some currency"
                                            },
                                            "payment_system": {
                                                "type": "string",
                                                "example": "PayPall"
                                            },
                                            "type": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "status": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}